refactor: added subcomponents to separate files
