+++ src/Dropdown/Dropdown.stories.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import Dropdown from './index';
+
 import Icon from '../Icon';
 import README from './README.md';
